+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Placeholder, necessary for Mypy to work properly."""